pyproject.toml
[tool.poetry]
name = "niche"
version = "0.1.0"
description = "Blog niche finder using crewAI"
authors = ["Sakhi Mehdi <mehdi.sakhi@proton.me>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = ">=0.55.2,<1.0.0" }
pytrends = { git = "https://github.com/MS-CICADA/pytrends-urllib3.git" }
exa_py = "1.1.0"
tavily-python = "0.5.0"
firecrawl-py = "1.2.4"
requests = "2.31.0"
python-dotenv = "1.0.0"
langchain-community = "0.2.17"


[tool.poetry.scripts]
run_crew = "niche.main:run"
train = "niche.main:train"
replay = "niche.main:replay"
test = "niche.main:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"